pyproject.toml
[project]
name = "SPN-Benchmarks"
version = "0.1.0"
description = "A project for generating and benchmarking Stochastic Petri Nets."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    # Core scientific and data handling packages
    "numpy==2.2.6",
    "scipy==1.16.1",
    "pandas",
    "matplotlib",
    "seaborn",
    "tabulate",
    "h5py==3.14.0",
    "joblib==1.5.2",
    "numba==0.61.2",
    "toml",

    # Utilities
    "tqdm==4.67.1",
    "graphviz==0.21",
    "scikit-learn==1.7.1",
    "jsonpickle==4.1.1",
    "jsonschema==4.25.1",

    # Excel file handling
    "openpyxl",

    # Cloud Storage
    "google-cloud-storage",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "pytest-benchmark",
    "memory-profiler",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120
target-version = ["py311"]